Deprecate Json fallback
