+++ app/src/main/java/com/cornellappdev/android/eatery/loginviews/LoginFragment.java
@@ -21,7 +21,7 @@ import com.cornellappdev.android.eatery.Repository;
 import com.cornellappdev.android.eatery.model.BrbInfoModel;
 import com.cornellappdev.android.eatery.model.enums.CacheType;
 import com.cornellappdev.android.eatery.network.GetLoginUtilities;
-import com.cornellappdev.android.eatery.network.JsonUtilities;
+import com.cornellappdev.android.eatery.network.QueryUtilities;
 import com.cornellappdev.android.eatery.presenter.AccountPresenter;
 import com.cornellappdev.android.eatery.util.InternalStorage;
 import com.google.firebase.analytics.FirebaseAnalytics;
@@ -109,7 +109,7 @@ public class LoginFragment extends Fragment {
 
                 @Override
                 public void successLogin(BrbInfoQuery.AccountInfo accountInfo) {
-                    Repository.getInstance().setBrbInfoModel(JsonUtilities.parseBrbInfo(accountInfo));
+                    Repository.getInstance().setBrbInfoModel(QueryUtilities.parseBrbInfo(accountInfo));
                     BrbInfoModel model = Repository.getInstance().getBrbInfoModel();
                     try {
                         InternalStorage.writeObject(getContext(), CacheType.BRB, model);
