--- app/src/main/java/com/cornellappdev/android/eatery/network/ConnectionUtilities.java
@@ -1,20 +1,0 @@
-package com.cornellappdev.android.eatery.network;
-
-import android.content.Context;
-import android.net.ConnectivityManager;
-import android.net.NetworkInfo;
-
-public class ConnectionUtilities {
-    public Context mContext;
-
-    public ConnectionUtilities(Context context) {
-        mContext = context;
-    }
-
-    public boolean isNetworkAvailable() {
-        ConnectivityManager connectivityManager =
-                (ConnectivityManager) mContext.getSystemService(Context.CONNECTIVITY_SERVICE);
-        NetworkInfo activeNetworkInfo = connectivityManager.getActiveNetworkInfo();
-        return activeNetworkInfo != null && activeNetworkInfo.isConnected();
-    }
-}
